refactoring
moved origin server response logic from HttpServerResponseConnection to HttpConnection
as this was only complicating logic.
Removed unused code.
Changed compile settings to Level 4

+++ BoostSocks5Server/SocksBindingServer.h
@@ -5,7 +5,7 @@ class SocksBindingServer
 {
 public:
  
-	SocksBindingServer(ba::io_service& ioService, unsigned int cAddress, unsigned short cPort, int bindingPort);
+	SocksBindingServer(ba::io_service& ioService, unsigned int cAddress, unsigned short cPort, unsigned short bindingPort);
     void StartAccept();
  
 private:
